pyproject.toml
[project]
name = "github-trending-newsletter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.13.4",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "idna==3.10",
    "jinja2==3.1.6",
    "lxml==5.4.0",
    "markupsafe==2.1.5",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "soupsieve==2.7",
    "typing-extensions==4.13.2",
    "urllib3==2.2.3",
]
